Update cryptoapp-min.js
Update the minified version of the main script for the Cryptocurrency Tracker.
Create indication for increase or decrease in the value of cryptocurrencies in a form of visual elements. Update card content generation logic with this new feature.
Implement logic for setting up price alert by pressing the "Enter" key when focused on the input field.
Create filtering functionality for cryptocurrencies based on their type according to user preferences.
Create sorting functionality for cryptocurrencies based on their parameters according to user preferences.
Update the URL for requests in such a way that allows to display 100 more cryptocurrencies.